restructured profile components into individual components
