+++ app/src/App.tsx
@@ -2,7 +2,7 @@ import "./App.css";
 import "./index.css";
 import Menu from "./Pages/Menu";
 import Login from "./Pages/Login";
-import Profile from "./components/Profile/Profile";
+import Profile from "./Pages/Profile";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Pong from "./components/Pong/Pong";
 import { useState } from "react";
