+++ app/src/components/Profile/ProfileContent.module.css
@@ -1,99 +1,3 @@
-/* * {
-    box-sizing: border-box;
-    margin: 0;
-    padding:0;
-} */
-
-.profile-container {
-    border-style: solid;
-    border-radius: 30px;
-    width: 80vw;
-    min-width: 500px;
-    max-width: 1800px;
-    height: 80vh;
-    min-height: 500px;
-    display: grid;
-    grid-template-rows: 200px auto;
-}
-
-/* Profile header */
-.profile-header-container {
-    border-style: solid;
-    border-color: red;
-    height: 100%;
-    width: 100%;
-    display: flex;
-}
-
-.profile-header-container > * {
-    border-style: solid;
-    border-color: pink;
-    flex: 1 1 0px;
-}
-
-.profile-id-container {
-    display: flex;
-    flex-direction: row;
-    flex-grow: 0;
-}
-
-.id-container-info {
-    display: flex;
-    flex-direction: column;
-    padding: 1rem;
-}
-
-.id-container-buttons {
-    display: flex;
-    flex-direction: column;
-    padding: 1rem 1rem 1rem 0;
-    row-gap: 1rem;
-}
-
-.id-container-buttons button {
-    width: 100px;
-}
-
-.profile-image {
-    border-radius: 50%;
-    border-style: solid;
-    border-color: green;
-    width: 100px;
-    min-height: 100px;
-    margin: auto;
-    margin-bottom: 1rem;
-}
-
-.profile-name-text {
-    margin: auto;
-    font-size: 1.25rem;
-}
-
-.profile-status {
-    margin: auto;
-    font-size: 1rem;
-}
-
-.primary-stats {
-    border-style: solid;
-    border-color: blue;
-    padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    min-height: 100%;
-}
-
-.primary-stats h3 {
-    border-style: solid;
-    border-color: green;
-    white-space: nowrap;
-    font-size: 2rem;
-}
-
-.secondary-stats p {
-    font-size: 1rem;
-}
-
 /* Profile Content */
 .profile-selector-container {
     border-style: solid;
